// packages/vercelai/src/types.ts
export interface MessageImageContentProps {
  src?: string;
  width?: string | number;
  height?: string | number;
  alt?: string;
}

/**
 * Type of message direction
 */
export type MessageDirection = 'incoming' | 'outgoing' | 0 | 1;

/**
 * Type of message type
 */
export type MessageType = 'html' | 'text' | 'image' | 'custom';

/**
 * Type of message content
 */
export type MessagePayload =
  | string
  | Record<string, unknown>
  | MessageImageContentProps
  | ReactNode;
// ...
export type CustomFunctionContext<C> = {
  [key: string]: C;
};

/**
 * Custom function to get the context object.
 * 
 * To support custom function execution, you can pass the context object to the custom function (see {@link CustomFunctionContext}).
 * You can also define a custom function to get the context object.
 */
export type CustomFunctionContextCallback<C> = () => CustomFunctionContext<C>;
// ...
export type CustomFunctionOutputProps<R, D> = {
  type: string;
  name: string;
  args?: Record<string, unknown>;
  isIntermediate?: boolean;
  result: R;
  data?: D;
  customMessageCallback?: CustomMessageCallback;
};
// ...
export type CallbackFunctionProps = {
  functionName: string;
  functionArgs: Record<string, unknown>;
  functionContext?: CustomFunctionContext<unknown> | CustomFunctionContextCallback<unknown>;
  previousOutput?: CustomFunctionOutputProps<unknown, unknown>[];
};

/**
 * Callback function for custom functions. You can define your own callback function to execute custom functions.
 * 
 * @param props - The props of the callback function. See {@link CallbackFunctionProps} for more details.
 * @returns The output of the custom function. See {@link CustomFunctionOutputProps} for more details.
 */
export type CallbackFunction = (props: CallbackFunctionProps) =>
  | CustomFunctionOutputProps<unknown, unknown>
  | Promise<CustomFunctionOutputProps<unknown, unknown>>;

/**
 * Type of Custom functions, a dictionary of functions e.g. createMap, createPlot etc.
 * key is the name of the function, value is the function itself.
 *
 * The function should return a CustomFunctionOutputProps object, or a Promise of CustomFunctionOutputProps object if it is a async function.
 */
export type CustomFunctions = {
  [key: string]: {
    func: CallbackFunction;
    context?: CustomFunctionContext<unknown> | CustomFunctionContextCallback<unknown>;
    callbackMessage?: CustomMessageCallback;
  };
};

/**
 * Type of CustomFunctionCall
 * 
 * @param functionName - The name of the function.
 * @param functionArgs - The arguments of the function.
 * @param output - The output of the function execution.
 */
export type CustomFunctionCall = {
  /** the name of the function */
  functionName: string;
  /** the arguments of the function */
  functionArgs: Record<string, unknown>;
  /** the output of function execution */
  output: CustomFunctionOutputProps<unknown, unknown>
};

/**
 * Type of CustomMessageCallback
 *
 * @param customFunctionCall The custom function call
 */
export type CustomMessageCallback = (
  customFunctionCall: CustomFunctionCall
) => ReactNode | null;

/**
 * Type of StreamMessageCallback
 *
 * @param deltaMessage The delta message from the assistant
 * @param customMessage The custom message from the custom function
 * @param isCompleted The flag to indicate if the message is completed
 */
export type StreamMessageCallback = (props: {
  deltaMessage: string;
  customMessage?: MessagePayload;
  isCompleted?: boolean;
}) => void;

/**
 * Type of UserActionProps
 * 
 * @param role - The role of the user.
 * @param text - The text of the user action.
 */
export type UserActionProps = {
  role: string;
  text: string;
};
// ...
export type ProcessMessageProps = {
  textMessage?: string;
  imageMessage?: string;
  userActions?: UserActionProps[];
  streamMessageCallback: StreamMessageCallback;
  useTool?: boolean;
  message?: string;
};
// ...
export type RegisterFunctionCallingProps = {
  name: string;
  description: string;
  properties: {
    [key: string]: {
      type: string; // 'string' | 'number' | 'boolean' | 'array';
      description: string;
      items?: {
        type: string;
      };
    };
  };
  required: string[];
  callbackFunction: CallbackFunction;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  callbackFunctionContext?: CustomFunctionContext<any>;
  callbackMessage?: CustomMessageCallback;
};